move user sign up schema to authentication.py model file
